requirements.txt
docker
gql[requests]
requests
boto3
opencv-python-headless
apscheduler
python-dotenv